requirements.txt
Flask~=3.0.3
colorama~=0.4.6
waitress~=3.0.0
requests~=2.32.3
psycopg-binary~=3.1.18
psycopg~=3.1.18

SQLAlchemy~=2.0.28
APScheduler~=3.11.0
paho-mqtt~=2.1.0